Refactor components directory, move to @/client
