+++ src/app/repos/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Repos from '@/components/Repos'
+import Repos from '@/client/components/Pages/Auth/Repos'
 import { VERSION_NUMBER } from '@/config/constants'
 
 export default function ReposPage() {
