+++ src/client/components/Pages/Auth/Home/index.ts
